refactor(VisualizationPanel): extract sub-components to satisfy size ceiling
VisualizationPanel.tsx had grown to 1654 lines, breaching the 1500-line
frontend-component ceiling enforced by `scripts/check_code_quality.py`.
Split out three self-contained concerns; the panel is now 1342 lines
and remains the orchestration shell for the four diagram tabs +
overflow iframe.

Extractions:

* `components/InspectSearchField.tsx` — the custom-rendered Inspect
  text field + suggestions dropdown that replaces the unreliable
  native `<input list>` / `<datalist>` pair when the host subtree is
  portaled into a detached window.
* `components/DetachedPlaceholder.tsx` — the in-place stand-in shown
  inside the panel when a tab is open in a separate browser window
  (Focus / Reattach affordances).
* `hooks/useOverflowIframe.ts` — the entire iframe ↔ parent-window
  plumbing for the overflow-graph tab: `cs4g:overlay-ready`
  handshake, message routing for pin / filter / layer / node events,
  pin-popover state with outside-click + Escape dismissal, and the
  pin / filter broadcast effects. Exposes the iframe ref, popover ref
  + state, and the popover setters needed by the parent JSX.

No behaviour changes — Vitest suite (1250 tests) + code-quality gate
both green; standalone bundle rebuilt and all four parity layers pass.
